backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-multipart
python-pptx
pandas
openpyxl
pymupdf
tabula-py
beautifulsoup4
playwright
requests
google-generativeai
python-dotenv
googlesearch-python
openai
httpx
celery
redis
boto3
camelot-py[cv]
reportlab
pydantic-settings
email-validator
psycopg2-binary
